update goals by clicking on the link
