+++ src/index.js
@@ -3,9 +3,8 @@ import "./css/style.css";
 import "./css/customStyles.css";
 import generateImages from "./generateImages";
 import Goal from "./models/goal";
-import generateDefault from "./generateDefault";
+import showGoal from "./showGoal";
 generateImages;
-generateDefault;
 
 // document.onreadystatechange = function () {
 //   var state = document.readyState
@@ -23,15 +22,11 @@ generateDefault;
 //   }
 // }
 
-
 const openModal = document.getElementById("open-modal-btn");
 const closeModal = document.getElementsByClassName("close-modal");
 const addGoalBtn = document.getElementById("add-goal-btn");
-
-const goalForm = {
-  title: document.getElementById("goal-title"),
-  category: document.getElementById("goal-category"),
-};
+const sidebarLinks = Array.from(document.getElementById('sidebar-links').getElementsByTagName('li'));
+const goalsContainer = document.getElementById("goals-container");
 
 const modal = document.getElementById("modal");
 
@@ -46,12 +41,48 @@ for (let index = 0; index < closeModal.length; index++) {
   });
 }
 
+
+// -------------------------
+let goals = [];
+
+const goal1 = new Goal("Goal 1", "all");
+const goal2 = new Goal("Goal 2", "today");
+const goal3 = new Goal("Goal 3", "week");
+const goal4 = new Goal("Goal 4", "year");
+const goal5 = new Goal("Goal 5", "year");
+
+goals.push(goal1);
+goals.push(goal2);
+goals.push(goal3);
+goals.push(goal4);
+goals.push(goal5);
+
+goals.forEach(goal=>{
+  showGoal(goal)
+})
+
+sidebarLinks.forEach(link=>{
+
+  link.addEventListener('click',()=>{
+    goalsContainer.innerHTML = "";
+
+    let text = link.getElementsByTagName('h3')[0].innerHTML.split(' ')[0].toLowerCase()
+    let filteredGoals = text === 'all'  ?  goals : goals.filter(goal=>goal.category === text)
+    filteredGoals.forEach(goal=>{
+      showGoal(goal)
+    })
+  })
+})
+
+// --------------------------------
+
+const goalForm = {
+  title: document.getElementById("goal-title"),
+  category: document.getElementById("goal-category"),
+};
+
 addGoalBtn.addEventListener("click", () => {
   const goal = new Goal(goalForm.title.value, goalForm.category.value);
-  goal.create();
-  all.innerHTML = `${Goal.goalsAmount.all.total}/${Goal.goalsAmount.all.done}`;
-  year.innerHTML = `${Goal.goalsAmount.year.total}/${Goal.goalsAmount.year.done}`;
-  week.innerHTML = `${Goal.goalsAmount.week.total}/${Goal.goalsAmount.week.done}`;
-  today.innerHTML = `${Goal.goalsAmount.today.total}/${Goal.goalsAmount.today.done}`;
+  showGoal(goal)
   modal.classList.add("hidden");
 });
