+++ src/partials/homepage/sidebar.hbs
@@ -1,8 +1,8 @@
 <aside class="fixed h-screen w-96 pl-32 pt-24">
   <div class="sidebar_bg bg-gray-200 shadow-sm">
-    <ul>
+    <ul id='sidebar-links'>
       <li class='flex justify-between mb-5 text-xl courgetteFont text-gray-500 cursor-pointer font-semibold'>
-        <h3> All goals</h3>
+        <h3>All goals</h3>
         <span id='all' />
       </li>
       <li class='flex justify-between mb-5 text-xl courgetteFont text-gray-500 cursor-pointer font-semibold'>
